Filter UnsupportedServiceVersion warning
Add the following warning emitted in functional tests
to the filterwarnings.

  UnsupportedServiceVersion: Service placement has no discoverable
  version. The resulting Proxy object will only have direct passthrough
  REST capabilities.

The warning is emitted because there is not a service description
(pairs of a suppoerted version and a proxy class) for placement service
in openstacksdk.
So set the filter until the service description for placement service
is added in openstacksdk.

Change-Id: Ie3d4a122ccfe1a1022e7aa7604585a727489c5be
Closes-Bug: #1838807