Allow assertXmlEqual() to pass options to matchers.XMLMatches
matchers.XMLMatches supports three options:

  - allow_mixed_nodes
  - skip_empty_text_nodes
  - skip_values

However these are not accessible when using the matcher via
assertXmlEqual().  So extend the latter with an *options
dictionary to allow options to be passed through.

This will be used by another commit in the SEV series in the near
future which needs the allow_mixed_nodes option to be enabled to
compare XML fragments with elements in different orders.

blueprint: amd-sev-libvirt-support
Change-Id: I31c60771891c3a2eb06fbc780ba2d369465498ec