Provider Config File: Function to further validate and retrieve configs
This series implements the referenced blueprint to allow for specifying
custom resource provider traits and inventories via yaml config files.

This second commit includes deeper validation of the custom resources,
as well as a function to load, parse, validate, and return all provider
configs from a directory containing provider.yaml files. This new
function is intended to be consumed by the resource tracker in a later
patch.

Co-Author: Tony Su <tao.su@intel.com>
Author: Dustin Cowles <dustin.cowles@intel.com>
Change-Id: I2d15af56fece3c5bba7583d51578b57f76a5fb70
Blueprint: provider-config-file