Avoid timeout from service update notification tests
Since If32bca070185937ef83f689b7163d965a89ec10a some of the service
update notification tests are slow because setting a compute enabled or
disabled makes an RPC call to the given compute service. These tests
does not have the proper RPC setup so that call times out. Since the
notification is still generated the test passes. Waiting for this
timeout steals 16 minutes of test execution time unnecessary.

This patch stubs the RPC call as that is not necessary for the
notification testing and therefore speeds up the tests.

Change-Id: I0269c4bd3da0e58593811b11985deb237eefe59e