Support reverting migration / resize with bandwidth
When a migration is reverted the source host allocation held by the
migration is swapped back to the instance. If the instance is bandwidth
aware then also the allocation key of the port's binding:profile needs
to be updated to point back to the resource providers of the source host
provider tree. To be able to do that we have to re-calculate the request
group - resource provider mapping for the source host based on the
resource requests in the neutron ports of the instance and the resource
allocation of the instance on the source host. Alternatively we could
store such mapping in the MigrationContext during the move operation.

blueprint: support-move-ops-with-qos-ports

Change-Id: Ib50b6b02208f5bd2972de8a6f8f685c19745514c