Enhance SDK fixture for 0.34.0
As we work out the kinks in openstacksdk to smooth the road for nova's
usage thereof, and until openstacksdk can provide opaque fixtures that
it can keep up to date as its internals change, we've been needing to
stub things out incrementally.

This commit more aggressively mocks the piece of openstacksdk that
keeps changing under us. It effectively does what [1] does, but without
the get_endpoint() sanity check, which blows up in unit/functional test
because it tries to do real ksa auth.

[1] https://review.opendev.org/#/c/675135/

Change-Id: I1ef2ca51da2dc25026a2fbdcfb1ed5199aaa518b