Bump minimum ksa (3.16.0) and sdk (0.34.0)
A number of different efforts are going to need to make use of
openstacksdk 0.34.0 and keystoneauth1 3.16.0, so rather than bump the
minimum in all of those, bump it in one place.

Also, this gives us the opportunity to independently validate some of
the fixes (particularly in logging) we were expecting to affect nova.

Change-Id: I87d1dcd299f6547d5f3c3d77e219bf71aba1cff2