Fix use of mock.patch with new_callable=PropertyMock
mock.patch's new_callable attribute is supposed to be set to a
callable object which when called will generate the new object for the
patched object to return:

  https://docs.python.org/dev/library/unittest.mock.html#unittest.mock.patch

So when it is desired to return a PropertyMock object, new_callable
should be set to the class, not an instance of the class.

Alternatively if it is desired to set the return value within the
decorator, then the new attribute should be used instead of
new_callable.

So change existing tests to adhere to this principle.  Previously they
passed more or less by accident, due to instances of PropertyMock
themselves being callable even though they are supposed to be used as
properties rather than methods.

Change-Id: I707169384c5732b9ad9f3941a9334f96478d4aab