Avoid timeout from service update api unit tests
Since If32bca070185937ef83f689b7163d965a89ec10a some of the service
update api tests are slow because setting a compute enabled or
disabled makes an RPC call to the given compute service. These tests
does not have the proper RPC setup so that call times out. The compute
api only logs an error for the timeout so the tests wasn't failing.

This patch stubs the RPC call as that is not necessary for the
api testing and therefore speeds up the tests.

Change-Id: If0960de896a67fd6bfca230b8915a45cb7af99b7