Provide HW_CPU_X86_AMD_SEV trait when SEV is supported
Add code to init_host() in the libvirt driver which detects support
for AMD SEV in the compute host hardware and hypervisor, by checking
that:

   a) 'sev' parameter of the kvm-amd kernel module is enabled, and

   b) the <features> element returned from libvirt's
      virConnectGetDomainCapabilities API contains an

         <sev supported='yes'>

      element.  This second check is achieved by utilising the domain
      capability retrieval and parsing code which was added in
      I4c35b6a27db05349213429422ffe62adb09bd921.

When both checks pass, provide the standard trait HW_CPU_X86_AMD_SEV
(which was newly available in os-traits 0.13.0) on the resource
provider for the compute host.  This indicates the ability of the
compute host to boot instances with AMD SEV (Secure Encrypted
Virtualization) enabled.

Note that this is not the last of the detection code.  Future commits
will add support for the use of a new MEM_ENCRYPTION_CONTEXT resource
class in the compute host inventory which tracks how many guests with
encrypted memory can be run on a host concurrently.  Until that point,
SEV functionality cannot be used, and this trait is effectively
dormant.

blueprint: amd-sev-libvirt-support
Change-Id: I2b41f1cce0af8b9d36b74a4663aa4ff227e17cc6