Support migrating SRIOV port with bandwidth
For normal ports with resource request it is enough to recalculate the
provider mapping during cold migrations. But for ports with PCI request
(direct, macvtap) the PCI request also needs to be updated to point to
the PF that provides the bandwidth on the destiantion host. So that the
pci claim on the compute allocates the VF from the same PF that provides
the bandwidth resource.

blueprint: support-move-ops-with-qos-ports

Change-Id: I2f3434f06489d8b6cb80933bcb1ea1e841049ba5