libvirt: Fix service-wide pauses caused by un-proxied libvirt calls
We only explicitly proxy the initial libvirt connection call.
We rely on tpool.Proxy's autowrap feature to ensure that all subsequent
calls are also proxied.

However:

* The list of classes passed to autowrap was incomplete
* autowrap doesn't work when a call returns a list of objects

We fix the problem of the autowrap whitelist by using inspection on the
libvirt module. We fix the only currently known case of a list being
returned by wrapping the contents explicitly.

Resolves-bug: #1840912
Change-Id: I668643c836d46a25df46d4c99a973af5e50a39db