Document archive_deleted_rows return codes
The archive_deleted_rows command has a specific set of return
codes for both errors and flow control so we should document
those in the CLI guide.

A FIXME is left in the code because of the odd case of when
you could get 1 back meaning "we archived some stuff" or
"something really bad happened".

Change-Id: Id10bd286249ad68a841f2bfb3a3623b429b2b3cc