neutron: refactor nw info cache refresh out of associate_floating_ip
This is a simple refactor to associate_floating_ip which moves
the code that handles refreshing the network info cache on the
instance originally associated with the floating IP out of the
main associate_floating_ip method. This is mostly to make unit
testing of that refesh code simpler in isolation.

Change-Id: Ie5d2724e61052756dc7239f25b1fb6eb0df96378