Allow resizing server with port resource request
This patch adds functional coverage for such resize. No extra code
changes were needed on top of the migrate patches to make resize work.

Now that the nova code supports the resize this patch removes the
check from the API that rejected the operation.

Note that in the spec [1] we agreed not to introduce new microversion
for this change but treat it as a bugfix. The current change also makes
it possible to accept the resize of these servers with _any_
microversion.

[1] https://specs.openstack.org/openstack/nova-specs/specs/train/approved/support-move-ops-with-qos-ports.html#rest-api-impact

blueprint: support-move-ops-with-qos-ports

Change-Id: Id0ee10e8d323786f4d79c82e3f05b48e76bd0956