fix lxml compatibility issues
Some unit tests were performing string matches on xml data
with inputs that were generated using lxml. This is problematic
as while white space between element tags is important in xml
ordering of attributes within a tag is not. In the latest version
of lxml the ordering asserted in the test no longer matches the
order returned by lxml on python 3.6+.

This change updates the failing test to use the XMLMatcher
class to compare xml strings instead.

Closes-Bug: #1838666
Related-Bug: #1841667
Change-Id: I1649a850ccb9ac85d7a962936ffef51d573b6f78