Rename Claims resources to compute_node
The "resources" arg to a Claim object is actually a ComputeNode
object so calling it "resources" is confusing. The name actually
goes back to when the claims module was added in change
Ib0b856376417f140dc9ed114913f47a8a790b8a2.

This simply renames that resources arg to compute_node so it's
more clear to people reading the code what that thing is.

Change-Id: I6859b1b840590a5aad63bf59fd4a86ccc5ef6353