Handle VirtDriverNotReady in _cleanup_running_deleted_instances
Change I6e086e0a2bffe12fd78373b3bc6215221dabc77e made the
_cleanup_running_deleted_instances task run immediately on
start of the nova-compute service. For some compute drivers, like
ironic, the "hypervisor" might not yet be ready to list instances
which results in a traceback exception in the logs.

This simply handles the VirtDriverNotReady error and returns to
avoid the traceback.

Change-Id: I6220a59988502a9839645efa101d555c19d4b650
Closes-Bug: #1842081