Use microversion in put allocations in test_report_client
In some places test_report_client uses a raw client.put when
writing allocations. This defaults to the latest microversion.
This means that if the allocations format changes in the latest
microversion in a patch in placement, the nova functional tests
fail.

In this change we pin the microversion to the minimum one providing
the features used in the PUT.

Change-Id: Ia0fee6bb931770792b552ae32ef31f0a4cc466ee