libvirt: Start reporting PCPU inventory to placement
This requires a lot of hand holding for legacy configuration options
(hence the plethora of tests) but this is all required to not break
upgrades.

A new '[compute] cpu_dedicated_set' config option is added, including a
wall of text necessary to explain the upgrade implications of this
option. The 'vcpu_pin_set' option is deprecated, while the help text for
the 'reserved_host_cpus' option is updated to specify that it will be
ignored for the libvirt driver.

Part of blueprint cpu-resources

Change-Id: I491eb7abe1405b27c552cdef142c688a46e914f2
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>