resize: Add bw min service level check of source compute
During resize the RequestSpec goes from the dest compute to the
source compute and then back to the dest. The previous patch [1] added
service level check for the dest compute. However the source compute also
needs to be new enough so the RequestSpec is passed through it.

Please note that the functional coverage for this api change is in a
later patch [2].

[1] https://review.opendev.org/#/c/680394
[2] https://review.opendev.org/#/c/679019

blueprint: support-move-ops-with-qos-ports
Change-Id: Ib1a73f5e20b6f9a325d8b24d9253a18f2a46db1f