Remove Rocky compute compat checks for live migration with port bindings
The nova-compute RPC API version 35 was added in Rocky [1] to signal
computes that can support live migration with the neutron port bindings
extended API. The check in conductor was added in Rocky as well [2] to
only use port bindings when both the source and destination compute
were running at least version 35. Now that we're past supporting rolling
upgrades from Rocky compute services we can remove that check in conductor.

[1] I1ba00850bdd7c015c5f462f1d75a5b833a5e9af9
[2] Ie04ded7b0f4edbb98af92f2ce713cf712a83b383

Change-Id: I007a75ad6743909028c3b23f614c693cee9e4a93