VMware: disk_io_limits settings are not reflected when resize
When resize and resize-revirt excute, setting disk_io_limits of
target flavor. But using VMware driver, disk_io_limits does not
changed.

On VMware driver, setting disk_io_limits when attach disk.
But, original is not setting it when resize. So add force
detach/attach disk when resize. Specifically, the original
_resize_disk() didn't anything if flavor.root_gb has not
changed. This is because we need to apply the change of
disk_io_limits to the virtual machine anytime.

Change-Id: I98f00257c38d95423bc926f3aa6a7b39dc84012e
Closes-bug: #1841514